Chore: DetailPage에서 DetailComponent로 변경
