+++ src/components/contents/detail/DetailComponent.module.css
@@ -1,0 +1,40 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  /* justify-content: center; */
+  height: 100vh;
+}
+
+.imageContainer figcaption {
+  margin: 8px 0;
+  padding: 16px 10px;
+  text-align: center;
+}
+
+.imageContainer .title {
+  font-size: 2.2rem;
+}
+
+.imageContainer .info {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 4px;
+  margin: 10px 0;
+  font-size: 1.4rem;
+  font-weight: 400;
+}
+
+.imageContainer .description {
+  font-size: 1.4rem;
+  text-align: left;
+  line-height: 160%;
+}
+
+.bottomContainer {
+  display: flex;
+  justify-content: space-around;
+  width: 100%;
+  height: 56px;
+}
