--- src/components/contents/detail/DetailPage.module.css
@@ -1,7 +1,0 @@
-.container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  height: 100vh;
-}
